refactor settings and language files into /config (#223)
* refactor settings files into /config

* refactor Language and LanguageManage into /config/language

* refactor MyAppCompatActivity.kt into /config
