+++ app/src/androidTest/java/com/github/fribourgsdp/radio/SettingsBuilderTest.kt
@@ -1,5 +1,8 @@
 package com.github.fribourgsdp.radio
 
+import com.github.fribourgsdp.radio.config.Settings
+import com.github.fribourgsdp.radio.config.SettingsBuilder
+import com.github.fribourgsdp.radio.config.language.Language
 import org.junit.Assert
 import org.junit.Test
