+++ app/src/main/java/com/github/fribourgsdp/radio/FireBaseTestActivity.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import android.content.Intent
 import android.os.Bundle
+import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.data.Genre
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
