+++ app/src/main/java/com/github/fribourgsdp/radio/config/MyAppCompatActivity.kt
@@ -1,7 +1,8 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.config
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.github.fribourgsdp.radio.config.language.LanguageManager
 
 open class MyAppCompatActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
