+++ app/src/main/java/com/github/fribourgsdp/radio/config/language/Language.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.config.language
 
 enum class Language( val code: String, val s: String) {
     ENGLISH("en", "English"),
